reject requests if there is no rate budget available
