+++ ingest/rate_limit.go
@@ -1,0 +1,67 @@
+package ingest
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+	"strings"
+
+	"golang.org/x/time/rate"
+)
+
+var (
+	rateLimiters map[int]*rate.Limiter // org id -> rate limiter
+)
+
+func ConfigureRateLimits(limitStr string) error {
+	if len(limitStr) == 0 {
+		return nil
+	}
+
+	rateLimiters = make(map[int]*rate.Limiter)
+
+	limitsWithOrgs := strings.Split(limitStr, ";")
+	for _, limitWithOrg := range limitsWithOrgs {
+		limitWithOrgParts := strings.SplitN(limitWithOrg, ":", 2)
+		if len(limitWithOrgParts) != 2 {
+			return fmt.Errorf("Invalid limit configuration string: %q", limitWithOrg)
+		}
+
+		orgId, err := strconv.ParseInt(limitWithOrgParts[0], 10, 32)
+		if err != nil {
+			return fmt.Errorf("Unable to parse orgId from string: %q", limitWithOrgParts[0])
+		}
+
+		limit, err := strconv.ParseInt(limitWithOrgParts[1], 10, 32)
+		if err != nil {
+			return fmt.Errorf("Unable to parse rate limit from string: %q", limitWithOrgParts[1])
+		}
+
+		rateLimiters[int(orgId)] = rate.NewLimiter(rate.Limit(limit), int(limit))
+	}
+
+	return nil
+}
+
+func rateLimit(ctx context.Context, orgId, datapoints int) {
+	limiter, ok := rateLimiters[orgId]
+	if !ok {
+		return
+	}
+
+	// wait until we are allowed to publish the given number of datapoints
+	limiter.WaitN(ctx, datapoints)
+}
+
+func IsRateBudgetAvailable(ctx context.Context, orgId int) bool {
+	limiter, ok := rateLimiters[orgId]
+	if !ok {
+		return true
+	}
+
+	return limiter.Allow()
+}
+
+func UseRateLimit() bool {
+	return len(rateLimiters) > 0
+}
